requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
discord==2.2.3
discord.py==2.2.3
exceptiongroup==1.1.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
numpy==1.24.3
outcome==1.2.0
pandas==2.0.2
pdfdocument==4.0.0
Pillow==9.5.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
schedule==1.2.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==2.0.2
wsproto==1.2.0
yarl==1.9.2
